+++ tiger-rbel/src/main/java/de/gematik/rbellogger/writer/RbelJwtSerializer.java
@@ -22,6 +22,11 @@ public class RbelJwtSerializer implements RbelSerializer {
         return renderToString(node, rbelWriter).getBytes();
     }
 
+    @Override
+    public byte[] renderNode(RbelContentTreeNode node, RbelWriterInstance rbelWriter) {
+        return render(node, rbelWriter);
+    }
+
     public String renderToString(RbelContentTreeNode node, RbelWriterInstance rbelWriter) {
         final JsonWebSignature jws = new JsonWebSignature();