+++ tiger-rbel/src/main/java/de/gematik/rbellogger/writer/RbelSerializer.java
@@ -10,4 +10,6 @@ import de.gematik.rbellogger.writer.tree.RbelContentTreeNode;
 public interface RbelSerializer {
     byte[] render(RbelContentTreeNode treeRootNode, RbelWriterInstance rbelWriter);
 
+    byte[] renderNode(RbelContentTreeNode treeRootNode, RbelWriterInstance rbelWriter);
+
 }